# docs/api/openapi.yaml
...
openapi: "3.0.3"
info:
  title: "Brookrege API"
  version: "1.0.0"
  description: "REST API of the Brookrege real estate platform (Sohag, Egypt).\n\n**Public** endpoints (`/api/...`) serve the website: no sign-in, rate-limited, responses cached for up to 30 minutes and cleared on every admin change.\n\n**Admin** endpoints (`/api/admin/...`) are served only on the admin domain and need a signed-in staff session: HttpOnly cookies set by `/auth/login` (+ `/auth/2fa/verify`). Requests must come from the admin app's origin (Origin / Sec-Fetch-Site checked), and may be limited to allowed networks. Each endpoint lists the permission it needs; roles: SUPER_ADMIN (all), CONTENT_ADMIN (listings, catalog, media, leads, analytics), MODERATOR (leads, listing lifecycle).\n\n**Errors** always have the shape `{ \"error\": { \"code\", \"message\", \"details\"? } }` with a plain-English message.\n\n**Prices** are EGP. Written content is Arabic first (`title`), English optional (`titleEn`); place names are Latin (`name`) with an Arabic form (`nameAr`)."
  contact:
    name: "Brookrege technical team"
    email: "security@brookrege.com"
servers:
  - url: "https://brookrege.com"
    description: "Public site"
  - url: "https://admin.brookrege.com"
    description: "Admin (only /api/admin/*)"
  - url: "http://localhost:4000"
    description: "Local development"
security:
  - session: []
...
paths:
...
  /api/admin/catalog/compounds:
    get:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "List compounds"
      operationId: "get_api_admin_catalog_compounds"
      description: "**Permission:** `property:read`."
      x-permission: "property:read"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "array"
                    items:
                      "$ref": "#/components/schemas/Json"
                required:
                  - "data"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
    post:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "Create one of the compounds"
      operationId: "post_api_admin_catalog_compounds"
      description: "**Permission:** `catalog:write`."
      x-permission: "catalog:write"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                name:
                  type: "string"
                  minLength: 2
                  maxLength: 120
                nameAr:
                  type: "string"
                  nullable: true
                developerName:
                  type: "string"
                  nullable: true
                description:
                  type: "string"
                  nullable: true
                descriptionEn:
                  type: "string"
                  nullable: true
                regionId:
                  type: "string"
                latitude:
                  type: "number"
                  nullable: true
                longitude:
                  type: "number"
                  nullable: true
                coverImageUrl:
                  type: "string"
                  nullable: true
                  format: "uri"
                isPublished:
                  type: "boolean"
                sortOrder:
                  type: "integer"
              required:
                - "name"
                - "regionId"
      responses:
        201:
          description: "Created"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    "$ref": "#/components/schemas/Json"
                required:
                  - "data"
        400:
          "$ref": "#/components/responses/E400"
        409:
          "$ref": "#/components/responses/E409"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
  /api/admin/catalog/compounds/{id}:
    patch:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "Edit one of the compounds"
      operationId: "patch_api_admin_catalog_compounds_id"
      description: "**Permission:** `catalog:write`."
      x-permission: "catalog:write"
      parameters:
        - "$ref": "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: "object"
                  properties:
                    name:
                      type: "string"
                      minLength: 2
                      maxLength: 120
                    nameAr:
                      type: "string"
                      nullable: true
                    developerName:
                      type: "string"
                      nullable: true
                    description:
                      type: "string"
                      nullable: true
                    descriptionEn:
                      type: "string"
                      nullable: true
                    regionId:
                      type: "string"
                    latitude:
                      type: "number"
                      nullable: true
                    longitude:
                      type: "number"
                      nullable: true
                    coverImageUrl:
                      type: "string"
                      nullable: true
                      format: "uri"
                    isPublished:
                      type: "boolean"
                    sortOrder:
                      type: "integer"
                  required:
                    - "name"
                    - "regionId"
              description: "Any subset of the fields."
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    "$ref": "#/components/schemas/Json"
                required:
                  - "data"
        400:
          "$ref": "#/components/responses/E400"
        404:
          "$ref": "#/components/responses/E404"
        409:
          "$ref": "#/components/responses/E409"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
    delete:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "Delete one of the compounds"
      operationId: "delete_api_admin_catalog_compounds_id"
      description: "**Permission:** `catalog:write`."
      x-permission: "catalog:write"
      parameters:
        - "$ref": "#/components/parameters/id"
      responses:
        204:
          description: "Done (no content)"
        404:
          "$ref": "#/components/responses/E404"
        409:
          "$ref": "#/components/responses/E409"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
  /api/admin/catalog/projects:
    get:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "List projects"
      operationId: "get_api_admin_catalog_projects"
      description: "**Permission:** `property:read`."
      x-permission: "property:read"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "array"
                    items:
                      "$ref": "#/components/schemas/Json"
                required:
                  - "data"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
    post:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "Create one of the projects"
      operationId: "post_api_admin_catalog_projects"
      description: "**Permission:** `catalog:write`."
      x-permission: "catalog:write"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                kind:
                  type: "string"
                  enum:
                    - "PARTNERSHIP"
                    - "COMPLETED"
                name:
                  type: "string"
                  minLength: 2
                  maxLength: 160
                nameAr:
                  type: "string"
                  nullable: true
                description:
                  type: "string"
                  nullable: true
                descriptionEn:
                  type: "string"
                  nullable: true
                partnerName:
                  type: "string"
                  nullable: true
                regionId:
                  type: "string"
                  nullable: true
                coverImageUrl:
                  type: "string"
                  nullable: true
                  format: "uri"
                completedAt:
                  type: "string"
                  format: "date-time"
                  nullable: true
                isPublished:
                  type: "boolean"
                sortOrder:
                  type: "integer"
              required:
                - "kind"
                - "name"
      responses:
        201:
          description: "Created"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    "$ref": "#/components/schemas/Json"
                required:
                  - "data"
        400:
          "$ref": "#/components/responses/E400"
        409:
          "$ref": "#/components/responses/E409"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
  /api/admin/catalog/projects/{id}:
    patch:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "Edit one of the projects"
      operationId: "patch_api_admin_catalog_projects_id"
      description: "**Permission:** `catalog:write`."
      x-permission: "catalog:write"
      parameters:
        - "$ref": "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: "object"
                  properties:
                    kind:
                      type: "string"
                      enum:
                        - "PARTNERSHIP"
                        - "COMPLETED"
                    name:
                      type: "string"
                      minLength: 2
                      maxLength: 160
                    nameAr:
                      type: "string"
                      nullable: true
                    description:
                      type: "string"
                      nullable: true
                    descriptionEn:
                      type: "string"
                      nullable: true
                    partnerName:
                      type: "string"
                      nullable: true
                    regionId:
                      type: "string"
                      nullable: true
                    coverImageUrl:
                      type: "string"
                      nullable: true
                      format: "uri"
                    completedAt:
                      type: "string"
                      format: "date-time"
                      nullable: true
                    isPublished:
                      type: "boolean"
                    sortOrder:
                      type: "integer"
                  required:
                    - "kind"
                    - "name"
              description: "Any subset of the fields."
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    "$ref": "#/components/schemas/Json"
                required:
                  - "data"
        400:
          "$ref": "#/components/responses/E400"
        404:
          "$ref": "#/components/responses/E404"
        409:
          "$ref": "#/components/responses/E409"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
    delete:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "Delete one of the projects"
      operationId: "delete_api_admin_catalog_projects_id"
      description: "**Permission:** `catalog:write`."
      x-permission: "catalog:write"
      parameters:
        - "$ref": "#/components/parameters/id"
      responses:
        204:
          description: "Done (no content)"
        404:
          "$ref": "#/components/responses/E404"
        409:
          "$ref": "#/components/responses/E409"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
  /api/admin/catalog/regions:
    get:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "List regions"
      operationId: "get_api_admin_catalog_regions"
      description: "**Permission:** `property:read`."
      x-permission: "property:read"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "array"
                    items:
                      "$ref": "#/components/schemas/Json"
                required:
                  - "data"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
    post:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "Create one of the regions"
      operationId: "post_api_admin_catalog_regions"
      description: "**Permission:** `catalog:write`."
      x-permission: "catalog:write"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                name:
                  type: "string"
                  minLength: 2
                  maxLength: 80
                nameAr:
                  type: "string"
                  nullable: true
                  maxLength: 80
                sortOrder:
                  type: "integer"
                isActive:
                  type: "boolean"
              required:
                - "name"
      responses:
        201:
          description: "Created"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    "$ref": "#/components/schemas/Json"
                required:
                  - "data"
        400:
          "$ref": "#/components/responses/E400"
        409:
          "$ref": "#/components/responses/E409"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
  /api/admin/catalog/regions/{id}:
    patch:
      tags:
        - "Admin — regions, compounds & projects"
      summary: "Edit one of the regions"
      operationId: "patch_api_admin_catalog_regions_id"
      description: "**Permission:** `catalog:write`."
      x-permission: "catalog:write"
      parameters:
        - "$ref": "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - type: "object"
                  properties:
                    name:
                      type: "string"
                      minLength: 2
                      maxLength: 80
                    nameAr:
                      type: "string"
                      nullable: true
                      maxLength: 80
                    sortOrder:
                      type: "integer"
                    isActive:
                      type: "boolean"
                  required:
                    - "name"
              description: "Any subset of the fields."
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    "$ref": "#/components/schemas/Json"
                required:
                  - "data"
        400:
          "$ref": "#/components/responses/E400"
        404:
          "$ref": "#/components/responses/E404"
        409:
          "$ref": "#/components/responses/E409"
        401:
          "$ref": "#/components/responses/E401"
        403:
          "$ref": "#/components/responses/E403"
...
  /api/compounds:
    get:
      tags:
        - "Public — catalog"
      summary: "Compounds with available units and starting prices"
      operationId: "get_api_compounds"
      security: []
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "array"
                    items:
                      "$ref": "#/components/schemas/CompoundSummary"
                required:
                  - "data"
        400:
          "$ref": "#/components/responses/E400"
  /api/compounds/{slug}:
    get:
      tags:
        - "Public — catalog"
      summary: "One compound"
      operationId: "get_api_compounds_slug"
      security: []
      parameters:
        - "$ref": "#/components/parameters/slug"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    "$ref": "#/components/schemas/Compound"
                required:
                  - "data"
        400:
          "$ref": "#/components/responses/E400"
        404:
          "$ref": "#/components/responses/E404"
...
  /api/projects:
    get:
      tags:
        - "Public — catalog"
      summary: "Partnership and completed projects"
      operationId: "get_api_projects"
      security: []
      parameters:
        - name: "kind"
          in: "query"
          schema:
            type: "string"
            enum:
              - "PARTNERSHIP"
              - "COMPLETED"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "array"
                    items:
                      "$ref": "#/components/schemas/Project"
                required:
                  - "data"
        400:
          "$ref": "#/components/responses/E400"
...
  /api/regions:
    get:
      tags:
        - "Public — catalog"
      summary: "Regions"
      operationId: "get_api_regions"
      security: []
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "array"
                    items:
                      "$ref": "#/components/schemas/Region"
                required:
                  - "data"
        400:
          "$ref": "#/components/responses/E400"
...
components:
  securitySchemes:
    session:
      type: "apiKey"
      in: "cookie"
      name: "__Host-bk_at"
      description: "Short-lived access cookie (15 min), renewed with /api/admin/auth/refresh. Plain-HTTP local development uses bk_at."
  parameters:
    id:
      name: "id"
      in: "path"
      required: true
      schema:
        type: "string"
        maxLength: 40
...
    slug:
      name: "slug"
      in: "path"
      required: true
      schema:
        type: "string"
        maxLength: 80
...
  responses:
    E400:
      description: "Invalid input (field messages in error.details)"
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    E401:
      description: "Not signed in, or the session ended"
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    E403:
      description: "Not allowed (role, origin, IP allowlist or a required step: code PASSWORD_CHANGE_REQUIRED / MFA_SETUP_REQUIRED)"
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    E404:
      description: "Not found"
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
    E409:
      description: "Conflict (e.g. in use)"
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
...
  schemas:
    Error:
      type: "object"
      properties:
        error:
          type: "object"
          properties:
            code:
              type: "string"
              example: "BAD_REQUEST"
            message:
              type: "string"
              description: "Plain-English explanation, safe to show."
            details:
              type: "object"
              additionalProperties:
                type: "array"
                items:
                  type: "string"
              description: "Per-field messages for validation errors."
          required:
            - "code"
            - "message"
      required:
        - "error"
...
    Place:
      type: "object"
      properties:
        name:
          type: "string"
        nameAr:
          type: "string"
          nullable: true
        slug:
          type: "string"
...
    TypeSummary:
      type: "object"
      properties:
        type:
          type: "string"
          enum:
            - "APARTMENT"
            - "VILLA"
            - "TOWNHOUSE"
            - "STUDIO"
            - "SHOP"
            - "ADMINISTRATIVE"
            - "LAND"
        startingFrom:
          type: "number"
          nullable: true
        availableUnits:
          type: "integer"
    Region:
      type: "object"
      properties:
        id:
          type: "string"
        name:
          type: "string"
        nameAr:
          type: "string"
          nullable: true
        slug:
          type: "string"
    CompoundSummary:
      type: "object"
      properties:
        id:
          type: "string"
        name:
          type: "string"
        nameAr:
          type: "string"
          nullable: true
        slug:
          type: "string"
        developerName:
          type: "string"
          nullable: true
        coverImageUrl:
          type: "string"
          nullable: true
        region:
          "$ref": "#/components/schemas/Place"
        availableUnits:
          type: "integer"
        startingFrom:
          type: "number"
          nullable: true
        types:
          type: "array"
          items:
            "$ref": "#/components/schemas/TypeSummary"
    Compound:
      type: "object"
      properties:
        id:
          type: "string"
        name:
          type: "string"
        nameAr:
          type: "string"
          nullable: true
        slug:
          type: "string"
        developerName:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
        descriptionEn:
          type: "string"
          nullable: true
        coverImageUrl:
          type: "string"
          nullable: true
        latitude:
          type: "number"
          nullable: true
        longitude:
          type: "number"
          nullable: true
        region:
          "$ref": "#/components/schemas/Place"
    Project:
      type: "object"
      properties:
        id:
          type: "string"
        kind:
          type: "string"
          enum:
            - "PARTNERSHIP"
            - "COMPLETED"
        name:
          type: "string"
        nameAr:
          type: "string"
          nullable: true
        slug:
          type: "string"
        description:
          type: "string"
          nullable: true
        descriptionEn:
          type: "string"
          nullable: true
        partnerName:
          type: "string"
          nullable: true
        coverImageUrl:
          type: "string"
          nullable: true
        completedAt:
          type: "string"
          format: "date-time"
          nullable: true
        region:
          nullable: true
          type: "object"
          properties:
            name:
              type: "string"
            nameAr:
              type: "string"
              nullable: true
...
    Json:
      type: "object"
      description: "See the admin screen that uses it; shape documented in the code."